No TLS keygen if key is already present

+++ ROOT/opt/instance-tools/bin/entrypoint.sh
@@ -74,8 +74,9 @@ main() {
     touch /.provisioning
 
     # Generate the Jupyter certificate if run in SSH/Args Jupyter mode
-    if [[ "${generate_tls_cert}" = "true" ]]; then
-        if [ ! -f  /etc/openssl-san.cnf ] || ! grep -qi vast /etc/openssl-san.cnf; then
+    sleep 2
+    if [[ "${generate_tls_cert}" = "true" ]] && [[ ! -f /etc/instance.key && ! -f /etc/instance.crt ]]; then
+        if [ ! -f /etc/openssl-san.cnf ] || ! grep -qi vast /etc/openssl-san.cnf; then
             echo "Generating certificates"
             echo '[req]' > /etc/openssl-san.cnf;
             echo 'default_bits       = 2048' >> /etc/openssl-san.cnf;
@@ -99,6 +100,11 @@ main() {
             openssl req -newkey rsa:2048 -subj "/C=US/ST=CA/CN=jupyter.vast.ai/" -nodes -sha256 -keyout /etc/instance.key -out /etc/instance.csr -config /etc/openssl-san.cnf
             curl --header 'Content-Type: application/octet-stream' --data-binary @//etc/instance.csr -X POST "https://console.vast.ai/api/v0/sign_cert/?instance_id=${CONTAINER_ID:-${VAST_CONTAINERLABEL#C.}}" > /etc/instance.crt;
         fi
+    fi
+
+    # If there is no key present we should ensure supervisor is aware
+    if [[ ! -f /etc/instance.key || ! -f /etc/instance.crt ]]; then
+        export ENABLE_HTTPS=false
     fi
 
     # Now we run supervisord - Put it in the background so provisioning can be monitored in Instance Portal
